requirements.txt
# Ingestion
kafka-python
confluent-kafka
jsonschema

# Feature Engineering
pandas
numpy
scipy
pyarrow

# ML
scikit-learn
xgboost
mlflow

# API
fastapi
uvicorn[standard]
pydantic

# Alerting
httpx
python-dotenv

# Storage
minio

# Validation
pycountry

# Tests
pytest
pytest-cov

# Utilities
python-dateutil
faker
